requirements.txt
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
djangochannelsrestframework==1.2.0
djangochannelsrestframework==1.2.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djoser==2.2.2
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
drf-yasg==1.21.7